+++ mayan/apps/document_states/models.py
@@ -161,8 +161,14 @@ class Workflow(models.Model):
                 initial_state = self.get_initial_state()
                 if initial_state:
                     for action in initial_state.entry_actions.filter(enabled=True):
+                        context = workflow_instance.get_context()
+                        context.update(
+                            {
+                                'action': action
+                            }
+                        )
                         action.execute(
-                            context=workflow_instance.get_context(),
+                            context=context,
                             workflow_instance=workflow_instance
                         )
             except IntegrityError: