+++ mayan/apps/document_states/queues.py
@@ -12,12 +12,20 @@ queue_document_states_fast = CeleryQueue(
 )
 
 queue_document_states.add_task_type(
-    label=_('Launch all workflows'),
+    label=_('Launch all workflows for all documents'),
     dotted_path='mayan.apps.document_states.tasks.task_launch_all_workflows'
 )
 queue_document_states.add_task_type(
     label=_('Launch a workflow'),
     dotted_path='mayan.apps.document_states.tasks.task_launch_workflow'
+)
+queue_document_states.add_task_type(
+    label=_('Launch a workflow for a document'),
+    dotted_path='mayan.apps.document_states.tasks.task_launch_workflow_for'
+)
+queue_document_states.add_task_type(
+    label=_('Launch all workflows for a document'),
+    dotted_path='mayan.apps.document_states.tasks.task_launch_all_workflow_for'
 )
 queue_document_states_fast.add_task_type(
     label=_('Generate workflow previews'),